Fix database connection

+++ Lab01/Person.cs
@@ -7,11 +7,11 @@ using System.Drawing;
 using System.Windows.Media.Imaging;
 
 namespace Lab01
-{   
+{
     public class Person
     {
         public string Name { get; set; }
-        public float Age { get; set; }
+        public int Age { get; set; }
         public BitmapImage Picture { get; set; }
     }
 }
